Refactor CLI common functions
